@@ -22,11 +22,12 @@
         }
     },
     "require": {
-        "php": ">=5.4",
+        "php": ">=7.1",
         "ext-json": "*"
     },
     "require-dev": {
-        "phpunit/phpunit": "^4.8.35|^5.7|^6.0",
-        "squizlabs/php_codesniffer": "1.*"
+        "phpunit/phpunit": "^7",
+        "squizlabs/php_codesniffer": "1.*",
+        "vimeo/psalm": "^3.8"
     }
 }
