@@ -1,7 +1,7 @@
 {
     "name": "symfony/event-dispatcher",
     "type": "library",
-    "description": "Symfony EventDispatcher Component",
+    "description": "Provides tools that allow your application components to communicate with each other by dispatching events and listening to them",
     "keywords": [],
     "homepage": "https://symfony.com",
     "license": "MIT",
@@ -16,18 +16,26 @@
         }
     ],
     "require": {
-        "php": "^5.5.9|>=7.0.8"
+        "php": ">=7.1.3",
+        "symfony/event-dispatcher-contracts": "^1.1",
+        "symfony/polyfill-php80": "^1.16"
     },
     "require-dev": {
-        "symfony/dependency-injection": "~3.3|~4.0",
-        "symfony/expression-language": "~2.8|~3.0|~4.0",
-        "symfony/config": "~2.8|~3.0|~4.0",
-        "symfony/debug": "~3.4|~4.4",
-        "symfony/stopwatch": "~2.8|~3.0|~4.0",
-        "psr/log": "~1.0"
+        "symfony/dependency-injection": "^3.4|^4.0|^5.0",
+        "symfony/expression-language": "^3.4|^4.0|^5.0",
+        "symfony/config": "^3.4|^4.0|^5.0",
+        "symfony/error-handler": "~3.4|~4.4",
+        "symfony/http-foundation": "^3.4|^4.0|^5.0",
+        "symfony/service-contracts": "^1.1|^2",
+        "symfony/stopwatch": "^3.4|^4.0|^5.0",
+        "psr/log": "^1|^2|^3"
     },
     "conflict": {
-        "symfony/dependency-injection": "<3.3"
+        "symfony/dependency-injection": "<3.4"
+    },
+    "provide": {
+        "psr/event-dispatcher-implementation": "1.0",
+        "symfony/event-dispatcher-implementation": "1.1"
     },
     "suggest": {
         "symfony/dependency-injection": "",
