@@ -1,7 +1,7 @@
 {
     "name": "symfony/finder",
     "type": "library",
-    "description": "Symfony Finder Component",
+    "description": "Finds files and directories via an intuitive fluent interface",
     "keywords": [],
     "homepage": "https://symfony.com",
     "license": "MIT",
@@ -16,7 +16,8 @@
         }
     ],
     "require": {
-        "php": "^5.5.9|>=7.0.8"
+        "php": ">=7.1.3",
+        "symfony/polyfill-php80": "^1.16"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Finder\\": "" },
