@@ -1,7 +1,7 @@
 {
     "name": "symfony/http-foundation",
     "type": "library",
-    "description": "Symfony HttpFoundation Component",
+    "description": "Defines an object-oriented layer for the HTTP specification",
     "keywords": [],
     "homepage": "https://symfony.com",
     "license": "MIT",
@@ -16,12 +16,14 @@
         }
     ],
     "require": {
-        "php": "^5.5.9|>=7.0.8",
+        "php": ">=7.1.3",
+        "symfony/mime": "^4.3|^5.0",
         "symfony/polyfill-mbstring": "~1.1",
-        "symfony/polyfill-php70": "~1.6"
+        "symfony/polyfill-php80": "^1.16"
     },
     "require-dev": {
-        "symfony/expression-language": "~2.8|~3.0|~4.0"
+        "predis/predis": "~1.0",
+        "symfony/expression-language": "^3.4|^4.0|^5.0"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\HttpFoundation\\": "" },
