@@ -16,7 +16,10 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3"
+        "php": ">=7.1"
+    },
+    "provide": {
+        "ext-ctype": "*"
     },
     "autoload": {
         "psr-4": { "Symfony\\Polyfill\\Ctype\\": "" },
@@ -28,7 +31,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-main": "1.19-dev"
+            "dev-main": "1.27-dev"
         },
         "thanks": {
             "name": "symfony/polyfill",
