@@ -15,9 +15,9 @@
 
 namespace think;
 
-if (version_compare(PHP_VERSION, '7.0.0', '<')) {
+if (version_compare(PHP_VERSION, '7.1.0', '<')) {
     header("Content-type: text/html; charset=utf-8");
-    die('PHP 7.0.0 及以上版本系统才可运行~ ');
+    die('PHP 7.1.0 及以上版本系统才可运行~ ');
 }
 
 define('IF_PUBLIC', true);
