@@ -220,7 +220,11 @@ class Template
 
             // 页面缓存
             ob_start();
-            ob_implicit_flush(0);
+            if (version_compare(PHP_VERSION, '8.0', '>=')) {
+                ob_implicit_flush(false);
+            } else {
+                ob_implicit_flush(0);
+            }
 
             // 读取编译存储
             $this->storage->read($cacheFile, $this->data);
