@@ -184,7 +184,11 @@ class View
 
         // 页面缓存
         ob_start();
-        ob_implicit_flush(0);
+        if (version_compare(PHP_VERSION, '8.0', '>=')) {
+            ob_implicit_flush(false);
+        } else {
+            ob_implicit_flush(0);
+        }
 
         // 渲染输出
         try {
